New File Cache Wrapper for new Matrix-SDK infrastructure
