+++ .github/workflows/check-native.yml
@@ -27,8 +27,8 @@ jobs:
             package: acter
             needs_flutter: true
             clippy_extras: --tests
-          - name: media-cache-wrapper
-            package: matrix-sdk-store-media-cache-wrapper
+          - name: file-event-cache
+            package: matrix-sdk-store-file-event-cache
     steps:
       - uses: actions/checkout@v4
       - name: Install rust toolchain
@@ -52,8 +52,8 @@ jobs:
           - name: Core
             package: acter-core
             clippy_extras: --tests
-          - name: media-cache-wrapper
-            package: matrix-sdk-store-media-cache-wrapper
+          - name: file-event-cache
+            package: matrix-sdk-store-file-event-cache
     steps:
       - uses: actions/checkout@v4
       - name: Install rust toolchain
